False Discorvery Rate (FDR)

Compute the False Discovery Rate for a vector of p-values and alpha value.

Usage

FDR(pvalues = NULL, alpha = 0.95, dep = 1)

pvalue.FDR(pvalues = NULL, dep = 1)

Arguments

pvalues

Vector of p-values

alpha

Alpha value (level of significance).

dep

Parameter dependence test. By default dep = 1, direct dependence between tests.

Value

Return:

out.FDR=TRUE: If there are significative differences.

pv.FDR: p-value for False Discovery Rate test.

Details

FDR method is used for multiple hypothesis testing to correct problems of multiple contrasts.
If dep = 1, the tests are positively correlated, for example when many tests are the same contrast.
If dep < 1 the tests are negatively correlated.

References

Benjamini, Y., Yekutieli, D. (2001). The control of the false discovery rate in multiple testing under dependency. Annals of Statistics. 29 (4): 1165-1188. DOI:10.1214/aos/1013699998.

See also

Author

Febrero-Bande, M. and Oviedo de la Fuente, M.

Examples

 p=seq(1:50)/1000
 FDR(p)
#> [1] TRUE
 pvalue.FDR(p)
#> [1] 0.05
 FDR(p,alpha=0.9999)
#> [1] FALSE
 FDR(p,alpha=0.9)
#> [1] TRUE
 FDR(p,alpha=0.9,dep=-1)
#> [1] FALSE