PCvM statistic for the Functional Linear Model with scalar response

Projected Cramer-von Mises statistic (PCvM) for the Functional Linear Model with scalar response (FLM): \(Y=\big<X,\beta\big>+\varepsilon\).

Adot(X, inpr)

PCvM.statistic(X, residuals, p, Adot.vec)

Arguments

X

Functional covariate for the FLM. The object must be either in the class fdata or in the class fd. It is used to compute the matrix of inner products.

inpr

Matrix of inner products of X. Computed if not given.

residuals

Residuals of the estimated FLM.

p

Number of elements of the functional basis where the functional covariate is represented.

Adot.vec

Output from the Adot function (see Details). Computed if not given.

Value

For PCvM.statistic, the value of the statistic. For Adot, a suitable output to be used in the argument Adot.vec.

Details

In order to optimize the computation of the statistic, the critical parts of these two functions are coded in FORTRAN. The hardest part corresponds to the function Adot, which involves the computation of a symmetric matrix of dimension \(n\times n\) where each entry is a sum of \(n\) elements. As this matrix is symmetric, the order of the method can be reduced from \(O(n^3)\) to \(O\big(\frac{n^3-n^2}{2}\big)\). The memory requirement can also be reduced to \(O\big(\frac{n^2-n+2}{2}\big)\). The value of Adot is a vector of length \(\frac{n^2-n+2}{2}\) where the first element is the common diagonal element and the rest are the lower triangle entries of the matrix, sorted by rows (see Examples).

Note

No NA's are allowed in the functional covariate.

References

Escanciano, J. C. (2006). A consistent diagnostic test for regression models using projections. Econometric Theory, 22, 1030-1051. doi:10.1017/S0266466606060506

Garcia-Portugues, E., Gonzalez-Manteiga, W. and Febrero-Bande, M. (2014). A goodness--of--fit test for the functional linear model with scalar response. Journal of Computational and Graphical Statistics, 23(3), 761-778. doi:10.1080/10618600.2013.812519

See also

Author

Eduardo Garcia-Portugues. Please, report bugs and suggestions to eduardo.garcia.portugues@uc3m.es

Examples

# Functional process
X=rproc2fdata(n=10,t=seq(0,1,l=101))
# Adot
Adot.vec=Adot(X)

# Obtain the entire matrix Adot
Ad=diag(rep(Adot.vec[1],dim(X$data)[1]))
Ad[upper.tri(Ad,diag=FALSE)]=Adot.vec[-1]
Ad=t(Ad)
Ad=Ad+t(Ad)-diag(diag(Ad))
Ad
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 34.55752 22.90508 23.85634 22.92308 23.15394 23.13439 23.07498 23.71767
#>  [2,] 22.90508 34.55752 23.26019 22.16730 22.77631 22.35080 23.17326 23.37389
#>  [3,] 23.85634 23.26019 34.55752 22.73172 23.57825 22.85524 22.51290 22.90221
#>  [4,] 22.92308 22.16730 22.73172 34.55752 23.03048 23.50064 22.58652 23.13668
#>  [5,] 23.15394 22.77631 23.57825 23.03048 34.55752 22.88934 22.36299 22.53649
#>  [6,] 23.13439 22.35080 22.85524 23.50064 22.88934 34.55752 23.42913 24.08463
#>  [7,] 23.07498 23.17326 22.51290 22.58652 22.36299 23.42913 34.55752 23.47460
#>  [8,] 23.71767 23.37389 22.90221 23.13668 22.53649 24.08463 23.47460 34.55752
#>  [9,] 23.61059 23.10960 23.33052 22.39196 23.12782 22.09138 22.44054 23.11028
#> [10,] 22.87736 23.84587 23.18498 22.94231 23.62108 22.51176 23.79073 22.54194
#>           [,9]    [,10]
#>  [1,] 23.61059 22.87736
#>  [2,] 23.10960 23.84587
#>  [3,] 23.33052 23.18498
#>  [4,] 22.39196 22.94231
#>  [5,] 23.12782 23.62108
#>  [6,] 22.09138 22.51176
#>  [7,] 22.44054 23.79073
#>  [8,] 23.11028 22.54194
#>  [9,] 34.55752 22.71785
#> [10,] 22.71785 34.55752
# Statistic
PCvM.statistic(X,residuals=rnorm(10),p=5)
#> [1] 5.271903