Smoothing matrix by nonparametric methods

Provides the smoothing matrix S for the discretization points tt

S.LLR(tt, h, Ker = Ker.norm, w = NULL, cv = FALSE)

S.LPR(tt, h, p = 1, Ker = Ker.norm, w = NULL, cv = FALSE)

S.LCR(tt, h, Ker = Ker.norm, w = NULL, cv = FALSE)

S.KNN(tt, h = NULL, Ker = Ker.unif, w = NULL, cv = FALSE)

S.NW(tt, h = NULL, Ker = Ker.norm, w = NULL, cv = FALSE)

Arguments

tt

Vector of discretization points or distance matrix mdist

h

Smoothing parameter or bandwidth. In S.KNN, number of k-nearest neighbors.

Ker

Type of kernel used, by default normal kernel.

w

Optional case weights.

cv

If TRUE, cross-validation is done.

p

Polynomial degree. be passed by default to create.basis

Value

Return the smoothing matrix S.

S.LLR return the smoothing matrix by Local Linear Smoothing.

S.NW return the smoothing matrix by Nadaraya-Watson kernel estimator.

S.KNN return the smoothing matrix by k nearest neighbors estimator.

S.LPR return the smoothing matrix by Local Polynomial Regression Estimator.

S.LCR return the smoothing matrix by Cubic Polynomial Regression.

Details

Options:

Nadaraya-Watson kernel estimator (S.NW) with bandwidth parameter h.

Local Linear Smoothing (S.LLR) with bandwidth parameter h.

K nearest neighbors estimator (S.KNN) with parameter knn.

Polynomial Local Regression Estimator (S.LCR) with parameter of polynomial p and of kernel Ker.

Local Cubic Regression Estimator (S.LPR) with kernel Ker.

References

Ferraty, F. and Vieu, P. (2006). Nonparametric functional data analysis. Springer Series in Statistics, New York.

Wasserman, L. All of Nonparametric Statistics. Springer Texts in Statistics, 2006.

Opsomer, J. D., and Ruppert, D. (1997). Fitting a bivariate additive model by local polynomial regression. The Annals of Statistics, 25(1), 186-211.

See also

Author

Manuel Febrero-Bande, Manuel Oviedo de la Fuente manuel.oviedo@udc.es

Examples

if (FALSE) {
  tt=1:101
  S=S.LLR(tt,h=5)
  S2=S.LLR(tt,h=10,Ker=Ker.tri)
  S3=S.NW(tt,h=10,Ker=Ker.tri)
  S4=S.KNN(tt,h=5,Ker=Ker.tri)
  par(mfrow=c(2,3))
  image(S)
  image(S2)
  image(S3)
  image(S4)
  S5=S.LPR(tt,h=10,p=1, Ker=Ker.tri)
  S6=S.LCR(tt,h=10,Ker=Ker.tri)
  image(S5)
  image(S6)
}