Performance measures for regression and classification models

cat2meas and tab2meas calculate the measures for a multiclass classification model.
pred2meas calculates the measures for a regression model.

cat2meas(yobs, ypred, measure = "accuracy", cost = rep(1, nlevels(yobs)))

tab2meas(tab, measure = "accuracy", cost = rep(1, nrow(tab)))

pred.MSE(yobs, ypred)

pred.RMSE(yobs, ypred)

pred.MAE(yobs, ypred)

pred2meas(yobs, ypred, measure = "RMSE")

Arguments

yobs

A vector of the labels, true class or observed response. Can be numeric, character, or factor.

ypred

A vector of the predicted labels, predicted class or predicted response. Can be numeric, character, or factor.

measure

Type of measure, see details section.

cost

Cost value by class (only for input factors).

tab

Confusion matrix (Contingency table: observed class by rows, predicted class by columns).

Details

cat2meas compute \(tab=table(yobs,ypred)\) and calls tab2meas function.

tab2meas function computes the following measures (see measure argument) for a binary classification model:

accuracy the accuracy classification score

recall, sensitivity,TPrate \(R=TP/(TP+FN)\)

precision \(P=TP/(TP+FP)\)

specificity,TNrate \(TN/(TN+FP)\)

FPrate \(FP/(TN+FP)\)

FNrate \(FN/(TP+FN)\)

Fmeasure \(2/(1/R+1/P)\)

Gmean \(sqrt(R*TN/(TN+FP))\)

kappa the kappa index

cost \(sum(diag(tab)/rowSums(tab)*cost)/sum(cost)\)

IOU \(TP/(TP+FN+FP)\) mean of Intersection over Union

IOU4class \(TP/(TP+FN+FP)\) Intersection over Union by level

pred2meas function computes the following measures of error, usign the measure argument, for observed and predicted vectors:

MSE Mean squared error, \(\frac{\sum{(ypred- yobs)^2}}{n} \)

RMSE Root mean squared error \(\sqrt{\frac{\sum{(ypred- yobs)^2}}{n} }\)

MAE Mean Absolute Error, \(\frac{\sum |yobs - ypred|}{n}\)

See also