aemet data

Series of daily summaries of 73 spanish weather stations selected for the period 1980-2009. The dataset contains geographic information of each station and the average for the period 1980-2009 of daily temperature, daily precipitation and daily wind speed.

Format

Elements of aemet:
..$df: Data frame with information of each wheather station:

ind: Indicated weather station.

name: Station Name. 36 marked UTF-8 strings.

province:Province (region) of Spain. 36 marked UTF-8 strings

altitude: Altitude of the station (in meters).

year.ini: Start year.

year.end: End year.

longitude: x geographic coordinate of the station (in decimal degrees).

latitude: y geographic coordinate of the station (in decimal degrees).

The functional variables:

...$temp: mean curve of the average daily temperature for the period 1980-2009 (in degrees Celsius, marked with UTF-8 string). In leap years temperatures for February 28 and 29 were averaged.

...$wind.speed: mean curve of the average daily wind speed for the period 1980-2009 (in m/s).

...$logprec: mean curve of the log precipitation for the period 1980-2009 (in log mm). Negligible precipitation (less than 1 tenth of mm) is replaced by 0.05 and no precipitation (0.0 mm) is replaced by 0.01. Then the logarithm is applied.

Source

The data were obtained from the FTP of AEMET in 2009.

Details

Meteorological State Agency of Spain (AEMET), https://www.aemet.es/es/portada. Government of Spain.
It marks 36 UTF-8 string of names of stations and 3 UTF-8 string names of provinces through the function iconv.

Author

Manuel Febrero Bande, Manuel Oviedo de la Fuente manuel.oviedo@udc.es

Examples

if (FALSE) {
data(aemet)
names(aemet)
names(aemet$df)
class(aemet)<-c("ldata","list") # ldata object
lat <- ifelse(aemet$df$latitude<31,"red","blue")
plot(aemet,col=lat)
}