One–way anova model for functional data

One–way anova model for k independent samples of functional data. The function contrasts the null hypothesis of equality of mean functions of functional data based on the an asymptotic version of the anova F–test. $$H_0:\, m_1=\ldots=m_k$$

Usage

fanova.onefactor(
  object,
  group,
  nboot = 100,
  plot = FALSE,
  verbose = FALSE,
  ...
)

Arguments

object

functional response data. fdata class object with n curves.

group

a factor specifying the class for each curve.

nboot

number of bootstrap samples.

plot

if TRUE, plot the mean of each factor level and the results of test.

verbose

if TRUE, print intermediate results.

...

further arguments passed to or from other methods.

Value

Returns:

p-value: Probability of rejecting the null hypothesis H0 at a significance level.

stat: Statistic value of the test.

wm: Statistic values of bootstrap resamples.

Details

The function returns the p–value of test using one–way anova model over nboot runs.

Note

anova.onefactor deprecated.

References

Cuevas, A., Febrero, M., & Fraiman, R. (2004). An anova test for functional data. Computational statistics & data analysis, 47(1), 111-122.

See also

Author

Juan A. Cuesta-Albertos, Manuel Febrero-Bande, Manuel Oviedo de la Fuente
manuel.oviedo@udc.es

Examples

if (FALSE) { # \dontrun{
data(MCO)
grupo<-MCO$classintact
datos<-MCO$intact
res=fanova.onefactor(datos,grupo,nboot=50,plot=TRUE)
grupo <- MCO$classpermea
datos <- MCO$permea
res=fanova.onefactor(datos,grupo,nboot=50,plot=TRUE)
} # }