Functional Data Analysis and Utilities for Statistical Computing (fda.usc)

This devel version carries out exploratory and descriptive analysis of functional data exploring its most important features: such as depth measurements or functional outliers detection, among others.
It also helps to explain and model the relationship between a dependent variable and independent (regression models) and make predictions. Methods for supervised or unsupervised classification of a set of functional data regarding a feature of the data are also included. Finally, it can perform analysis of variance model (ANOVA) for functional data.

Details

Sections of fda.usc-package:

A.- Functional Data Representation

B.- Functional Outlier Detection

C.- Functional Regression Model

D.- Functional Supervised Classification

E.- Functional Non-Supervised Classification

F.- Functional ANOVA

G.- Auxiliary functions

A.- Functional Data Representation
The functions included in this section allow to define, transform, manipulate and represent a functional dataset in many ways including derivatives, non-parametric kernel methods or basis representation.

fdata

plot.fdata

fdata.deriv

CV.S

GCV.S

optim.np

optim.basis

S.NW

S.LLR

S.basis

Var.e

Var.y

B.- Functional Depth and Functional Outlier Detection

The functional data depth calculated by the different depth functions implemented that could be use as a measure of centrality or outlyingness.

B.1-Depth methods Depth:

depth.FM

depth.mode

depth.RP

depth.RT

depth.RPD

Descriptive

B.2-Functional Outliers detection methods:

outliers.depth.trim

outliers.depth.pond

outliers.thres.lrt

outliers.lrt

C.- Functional Regression Models

C.1. Functional explanatory covariate and scalar response
The functions included in this section allow the estimation of different functional regression models with a scalar response and a single functional explicative covariate.

fregre.pc

fregre.pc.cv

fregre.pls

fregre.pls.cv

fregre.basis

fregre.basis.cv

fregre.np

fregre.np.cv

C.2. Test for the functional linear model (FLM) with scalar response.

flm.Ftest, F-test for the FLM with scalar response

flm.test, Goodness-of-fit test for the FLM with scalar response

PCvM.statistic, PCvM statistic for the FLM with scalar response

C.3. Functional and non functional explanatory covariates.
The functions in this section extends those regression models in previous section in several ways.

fregre.plm: Semifunctional Partial Linear Regression (an extension of lm model)

fregre.lm: Functional Linear Regression (an extension of lm model)

fregre.glm: Functional Generalized Linear Regression (an extension of glm model)

fregre.gsam: Functional Generalized Spectral Additive Regression (an extension of gam model)

fregre.gkam: Functional Generalized Kernel Additive Reg. (an extension of fregre.np model)

C.4. Functional response model (fregre.basis.fr) allows the estimation of functional regression models with a functional response and a single functional explicative covariate.

C.5. fregre.gls fits functional linear model using generalized least squares. fregre.igls fits iteratively a functional linear model using generalized least squares.

C.6. fregre.gsam.vs, Variable Selection using Functional Additive Models

D.- Functional Supervised Classification
This section allows the estimation of the groups in a training set of functional data fdata class by different nonparametric methods of supervised classification. Once these classifiers have been trained, they can be used to predict on new functional data.

Package allows the estimation of the groups in a training set of functional data by different methods of supervised classification.

D.1 Univariate predictor (x,y arguments, fdata class)

classif.knn

classif.kernel

D.2 Multiple predictors (formula,data arguments, ldata class)

classif.glm

classif.gsam

classif.gkam

D.3 Depth classifiers (fdata or ldata class)

classif.DD

classif.depth

D.4 Functional Classification usign k-fold CV

classif.kfold

E.- Functional Non-Supervised Classification
This section allows the estimation of the groups in a functional data set fdata class by kmeans method.

kmeans.fd

F.- Functional ANOVA

fanova.onefactor

fanova.RPm

fanova.hetero

G.- Utilities and auxiliary functions:

fdata.bootstrap

fdata2fd

fdata2pc

fdata2pls

summary.fdata.comp

cond.F

cond.quantile

cond.mode

FDR

Kernel

Kernel.asymmetric

Kernel.integrate

metric.lp

metric.kl

metric.DTW

metric.hausdorff

metric.dist

semimetric.NPFDA

semimetric.basis

Package:	fda.usc

Type:	Package

Version:	2.0.3

Date:	2021-06-03

License:	GPL-2

LazyLoad:	yes

References

Febrero-Bande, M., Oviedo de la Fuente, M. (2012). Statistical Computing in Functional Data Analysis: The R Package fda.usc. Journal of Statistical Software, 51(4), 1-28. doi:10.18637/jss.v051.i04

Author

Authors: Manuel Febrero Bande manuel.febrero@usc.es and Manuel Oviedo de la Fuente manuel.oviedo@udc.es

Contributors: Pedro Galeano, Alicia Nieto-Reyes, Eduardo Garcia-Portugues and STAPH group https://www.math.univ-toulouse.fr/~ferraty/

Maintainer: Manuel Oviedo de la Fuente manuel.oviedo@udc.es