fda.usc internal functions

Internal undocumentation functions for fda.usc package.

Usage

trace.matrix(x, na.rm = TRUE)

argvals.equi(tt)

# S3 method for class 'fdata'
fdata1 + fdata2

# S3 method for class 'fdata'
fdata1 - fdata2

# S3 method for class 'fdata'
fdata1 * fdata2

# S3 method for class 'fdata'
fdata1/fdata2

# S3 method for class 'fdata'
fdataobj[i = TRUE, j = TRUE, drop = FALSE]

# S3 method for class 'fdata'
fdata1 != fdata2

# S3 method for class 'fdata'
fdata1 == fdata2

# S3 method for class 'fdata'
fdataobj^pot

# S3 method for class 'fdata'
dim(x)

ncol.fdata(x)

nrow.fdata(x)

# S3 method for class 'fdata'
length(x)

NROW.fdata(x)

NCOL.fdata(x)

rownames.fdata(x)

colnames.fdata(x)

# S3 method for class 'fdata'
c(...)

argvals(fdataobj)

rangeval(fdataobj)

# S3 method for class 'fdist'
fdataobj[i = TRUE, j = TRUE, drop = FALSE]

# S3 method for class 'fdata'
is.na(x)

# S3 method for class 'fdata'
anyNA(x, recursive = FALSE)

count.na.fdata(x)

unlist_fdata(x, recursive = TRUE, use.names = TRUE)

Arguments

x

matrix or fdata class object.

na.rm

logical. Should missing values (including NaN) be removed?

tt

Argvals

fdataobj, fdata1, fdata2

fdata class object.

i, j

Indices specifying elements to extract, replace. Indices are numeric or character vectors or empty

drop

For fdata class object. If TRUE the result is coerced to the lowest possible dimension of element data. This only works for extracting elements, not for the replacement.

pot

Numeric value for exponentiation.

...

fdata objects to be concatenated.

recursive

should anyNA be applied recursively to lists and pairlists?
(in anyNA.fdata function) logical Should unlisting be applied to list components of x? (in unlist_fdata function).

use.names

logical Should names be preserved?

Details

argvals.equi function returns TRUE if the argvals are equispaced and FALSE in othercase.

Note

In "Ops" functions "+.fdata", "-.fdata", "*.fdata" and "/.fdata": The lengths of the objects fdata1 and fdata2 may be different because operates recycled into minimum size as necessary.

References

Febrero-Bande, M., Oviedo de la Fuente, M. (2012). Statistical Computing in Functional Data Analysis: The R Package fda.usc. Journal of Statistical Software, 51(4), 1-28. doi:10.18637/jss.v051.i04