Functional influence measures

Once estimated the functional regression model with scalar response, influence.fregre.fd function is used to obtain the functional influence measures.

Usage

# S3 method for class 'fregre.fd'
influence(model, ...)

Arguments

model

fregre.pc, fregre.basis or fregre.basis.cv object.

...

Further arguments passed to or from other methods.

Value

Return:

DCP: Cook's Distance for Prediction.

DCE: Cook's Distance for Estimation.

DP:\(\mbox{Pe}\tilde{\mbox{n}}\mbox{a's} \) Distance.

Details

Identify influential observations in the functional linear model in which the predictor is functional and the response is scalar. Three statistics are introduced for measuring the influence: Distance Cook Prediction DCP, Distance Cook Estimation DCE and Distance \(\mbox{pe}\tilde{\mbox{n}}\mbox{a} \) DP respectively.

Note

influence.fdata deprecated.

References

Febrero-Bande, M., Galeano, P. and Gonzalez-Manteiga, W. (2010). Measures of influence for the functional linear model with scalar response. Journal of Multivariate Analysis 101, 327-339.

Febrero-Bande, M., Oviedo de la Fuente, M. (2012). Statistical Computing in Functional Data Analysis: The R Package fda.usc. Journal of Statistical Software, 51(4), 1-28. https://www.jstatsoft.org/v51/i04/

See also

Author

Manuel Febrero-Bande, Manuel Oviedo de la Fuente manuel.oviedo@udc.es

Examples

if (FALSE) { # \dontrun{
data(tecator)
x=tecator$absorp.fdata[1:129]
y=tecator$y$Fat[1:129]

res1=fregre.pc(x,y,1:5)  
# time consuming
res.infl1=influence(res1)  
res2=fregre.basis(x,y)  
res.infl2=influence(res2)  

res<-res1
res.infl<-res.infl1
mat=cbind(y,res$fitted.values,res.infl$DCP,res.infl$DCE,res.infl$DP)
colnames(mat)=c("Resp.","Pred.","DCP","DCE","DP")
pairs(mat)
} # }