Quantile for influence measures

Estimate the quantile of measures of influence for each observation.

influence_quan(model,out.influ,mue.boot=500,
smo=0.1,smoX=0.05,alpha=0.95,kmax.fix=FALSE,...)

Arguments

model

fregre.pc, fregre.basis or fregre.basis.cv object.

out.influ

inflluence.fd bject

mue.boot

Number of bootstrap samples

smo

Smoothing parameter as a proportion of response variance.

smoX

Smoothing parameter for fdata object as a proportion of variance-covariance matrix of the explanatory functional variable.

alpha

Significance level.

kmax.fix

The maximum number of principal comoponents or number of basis is fixed by model object.

...

Further arguments passed to or from other methods.

Value

Return:

quan.cook.for Distance Cook Prediction Quantile.

quan.cook.est Distance Cook Estimation Quantile.

quan.cook.Pena Pena Distance Quantile.

mues.est Sample Cook generated.

mues.pena Sample Pena generated.

beta.boot Functional beta estimated by bootstrap method.

Details

Compute the quantile of measures of influence estimated in influence.fregre.fd for functional regression using principal components representation (fregre.pc) or basis representation
(fregre.basis or fregre.basis.cv).

A smoothed bootstrap method is used to estimate the quantiles of the influence measures, which allows to point out which observations have the larger influence on estimation and prediction.

References

Febrero-Bande, M., Galeano, P. and Gonzalez-Manteiga, W. (2010). Measures of influence for the functional linear model with scalar response. Journal of Multivariate Analysis 101, 327-339.

See also

Author

Manuel Febrero-Bande, Manuel Oviedo de la Fuente manuel.oviedo@udc.es

Examples

if (FALSE) {
data(tecator)
x=tecator$absorp.fdata
y=tecator$y$Fat
res=fregre.pc(x,y,1:6)

#time consuming
res.infl=influence.fregre.fd(res)
resquan=influence_quan(res,res.infl,4,0.01,0.05,0.95)
plot(res.infl$betas,type="l",col=2)
lines(res$beta.est,type="l",col=3)
lines(resquan$betas.boot,type="l",col="gray")

res=fregre.basis(x,y)
res.infl=influence.fregre.fd(res)
resquan=influence_quan(res,res.infl,mue.boot=4,kmax.fix=T)
plot(resquan$betas.boot,type="l",col=4)
lines(res.infl$betas,type="l",col=2)
lines(resquan$betas.boot,type="l",col="gray")
}