Distance Matrix Computation

This function computes the distances between the rows of a data matrix by using the specified distance measure.

This function returns a distance matrix by using dist function.
The matrix dimension is (n1 x n1) if y=NULL, (n1 x n2) otherwise.

metric.dist(x, y = NULL, method = "euclidean", p = 2, dscale = 1, ...)

Arguments

x

Data frame 1. The dimension is (n1 x m).

y

Data frame 2. The dimension is (n2 x m).

method

The distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".

p

The power of the Minkowski distance.

dscale

If scale is a numeric, the distance matrix is divided by the scale value. If scale is a function (as the mean for example) the distance matrix is divided by the corresponding value from the output of the function.

...

Further arguments passed to dist function.

See also

Author

Manuel Febrero-Bande, Manuel Oviedo de la Fuente manuel.oviedo@udc.es

Examples

if (FALSE) {
data(iris)
d<-metric.dist(iris[,1:4])
matplot(d,type="l",col=as.numeric(iris[,5]))
}