Distance Matrix Computation for ldata and mfdata class object

This function computes the distances between the list elements

metric.ldata(
  ldata1,
  ldata2 = NULL,
  include = "all",
  exclude = "none",
  metric,
  par.metric = NULL,
  w,
  method = "none"
)

Arguments

ldata1

List with of fdata objects and a data.frame object calle 'df'.

ldata2

List with of fdata objects and a data.frame object calle 'df'.

include

vector with the name of variables to use

exclude

vector with the name of variables to not use

metric

Type of metric to combine, if 'none', the function no combine and return a list o distances for each variable included

par.metric

List of metric parameters for each variable included

w,

weights to combine the metric (if metric is not 'none')

method

The distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".

Details

This function returns a distance matrix by using metric.lp function for fdata objects and metric.dist function for vector and matric objects.

See also

Author

Manuel Febrero-Bande, Manuel Oviedo de la Fuente manul.oviedo@usc.es

Examples

if (FALSE) {
data(tecator)
names(tecator)[2] <- "df"
# Example 1 (list of distances)
ldist <- metric.ldata(tecator,method="none")
lapply(ldist,names)
# Example 2 (combined metric)
mdist <- metric.ldata(tecator,method="euclidean")
dim(mdist) 
}