ops.fda.usc Options Settings

Set or query graphical and prompt output parameters. Allow the user to set and examine a variety of global or local options which affect the way in which fda.usc functions computes and displays its results.

Usage

ops.fda.usc(
  verbose = FALSE,
  trace = FALSE,
  warning = FALSE,
  ncores = NULL,
  int.method = "TRAPZ",
  reset = FALSE,
  eps = as.double(.Machine[[1]] * 10)
)

Arguments

verbose

logical. Should R report extra information on progress? Set to TRUE by the command-line option –verbose.

trace

logical. Show internal information of procedure.

warning

logical: If true, warnings are shown.

ncores

integer. Number of CPU cores on the current host.

int.method

see method argument in int.simpson function.

reset

logical. If TRUE creates a new Parallel Socket Cluster (ncores>1) or a sequential parallel backend (ncores=1). It is useful when worker initialization failed or after a crush.

eps

epsilon parameter.

Author

Manuel Oviedo de la Fuente (manuel.oviedo@udc.es).

Examples

if (FALSE) { # \dontrun{
# If worker initialization failed, please execute this code
 ncores <- max(parallel::detectCores() -1,1)
 if (ncores==1) {
     foreach::registerDoSEQ()
 }  else{
 cl <-  suppressWarnings(parallel::makePSOCKcluster(ncores ))
 doParallel::registerDoParallel(cl)
 }
 ops.fda.usc()
} # }