Wild bootstrap residuals

The wild bootstrap residuals are computed as \(residuals*V\), where \(V\) is a sampling from a random variable (see details section).

rwild(residuals, type = "golden")

Arguments

residuals

residuals

type

Type of distribution of V.

Value

The wild bootstrap residuals computed using a sample of the random variable \(V\).

Details

For the construction of wild bootstrap residuals, sampling from a random variable \(V\) such that \(E[V^2]=0\) and \(E[V]=0\) is needed. A simple and suitable \(V\) is obtained with a discrete variable of the form:

``golden'', Sampling from golden section bootstrap values suggested by Mammen (1993). $$P\Bigg\{ V=\frac{1-\sqrt{5}}{2} \Bigg\} = \frac{5+\sqrt{5}}{10} \, and \, P\Bigg\{ V=\frac{1+\sqrt{5}}{2} \Bigg\} = \frac{5-\sqrt{5}}{10},$$ which leads to the golden section bootstrap.

``Rademacher'', Sampling from Rademacher distribution values \(\big\{-1,\,1\big\}\) with probabilities \(\big\{\frac{1}{2},\,\frac{1}{2}\big\}\), respectively.

``normal'', Sampling from a standard normal distribution.

References

Mammen, E. (1993). Bootstrap and wild bootstrap for high dimensional linear models. Annals of Statistics 21, 255 285. Davidson, R. and E. Flachaire (2001). The wild bootstrap, tamed at last. working paper IER1000, Queens University.

See also

Author

Eduardo Garcia-Portugues, Manuel Febrero-Bande and Manuel Oviedo de la Fuente manuel.oviedo@udc.es.

Examples

n<-100
# For golden wild bootstrap variable
e.boot0=rwild(rep(1,len=n),"golden")
# Construction of wild bootstrap residuals
e=rnorm(n)
e.boot1=rwild(e,"golden")
e.boot2=rwild(e,"Rademacher")
e.boot3=rwild(e,"normal")
summary(e.boot1)
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#> -5.2594 -0.7925 -0.0571 -0.1984  0.5094  3.1499 
summary(e.boot2)
#>     Min.  1st Qu.   Median     Mean  3rd Qu.     Max. 
#> -3.71851 -0.75170  0.03608 -0.03810  0.85791  3.46545 
summary(e.boot3)
#>      Min.   1st Qu.    Median      Mean   3rd Qu.      Max. 
#> -3.435818 -0.437566 -0.003045 -0.095816  0.256373  3.797496