Summarizes information from fregre.gkam objects.

Summary function for fregre.gkam function.

-Family used.

-Number or iteration of algorithm and if it has converged.

-Residual and null deviance.

-Number of data.

Produces a list of summary information for a fitted fregre.np object for each functional covariate.

-Call.

-R squared.

-Residual variance.

-Index of possible atypical curves or possible outliers.

-Index of possible influence curves.

If draw=TRUE plot:

-y vs y fitted values.

-Residuals vs fitted values.

-Residual boxplot.

-Quantile-Quantile Plot (qqnorm).

-Plot for a each single model term.

If ask=FALSE draw graphs in one window, by default. If ask=TRUE, draw each graph in a window, waiting to confirm.

Usage

# S3 method for class 'fregre.gkam'
summary(object, draw = TRUE, selec = NULL, times.influ = 3, ...)

Arguments

object

Estimated by functional regression, fregre.fd object.

draw

=TRUE draw estimation and residuals graphics.

selec

Allows the plot for a single model term to be selected for printing. e.g. if you just want the plot for the second smooth term set selec=2. .

times.influ

Limit for detect possible infuence curves.

...

Further arguments passed to or from other methods.

See also

Author

Manuel Febrero-Bande and Manuel Oviedo de la Fuente manuel.oviedo@udc.es

Examples

if (FALSE) { # \dontrun{
# Time consuming
data(tecator)
ind<-1:129
ab=tecator$absorp.fdata[ind]
ab2=fdata.deriv(ab,2)
yfat=as.integer(cut(tecator$y[ind,"Fat"],c(0,15,100)))-1
xlist=list("df"=data.frame(yfat),"ab2"=ab2,"ab"=ab)
f<-yfat~ab+ab2
res=fregre.gkam(f, data = xlist, family = binomial("logit"),
                control = list(maxit = 2))
summary(res)

} # }