outliers for functional dataset

Procedure for detecting funcitonal outliers.

Usage

outliers.depth.pond(
  fdataobj,
  nb = 200,
  smo = 0.05,
  quan = 0.5,
  dfunc = depth.mode,
  ...
)

outliers.depth.trim(
  fdataobj,
  nb = 200,
  smo = 0.05,
  trim = 0.01,
  quan = 0.5,
  dfunc = depth.mode,
  ...
)

outliers.lrt(fdataobj, nb = 200, smo = 0.05, trim = 0.1, ...)

outliers.thres.lrt(fdataobj, nb = 200, smo = 0.05, trim = 0.1, ...)

Arguments

fdataobj

fdata class object.

nb

The number of bootstrap samples.

smo

The smoothing parameter for the bootstrap samples.

quan

Quantile to determine the cutoff from the Bootstrap procedure (by default=0.5)

dfunc

Type of depth measure, by default depth.mode.

...

Further arguments passed to or from other methods.

trim

The alpha of the trimming.

Value

outliers

Indexes of functional outlier.

dep.out

Depth value of functional outlier.

dep.out

Iteration in which the functional outlier is detected.

quantile

Threshold for outlier detection.

dep

Depth value of functional data.

Details

Outlier detection in functional data by likelihood ratio test (outliers.lrt). The threshold for outlier detection is given by the outliers.thres.lrt. Outlier detection in functional data by depth measures:

outliers.depth.pond: function weights the data according to depth.

outliers.depth.trim: function uses trimmed data.

quantile_outliers_pond and quantile_outliers_trim functions provides the quantiles of the bootstrap samples for functional outlier detection by, respectively, weigthed and trimmed procedures. Bootstrap smoothing function (fdata.bootstrap with nb resamples) is applied to these weighted or trimmed data. If smo=0 smoothed bootstrap is not performed. The function returns a vector of size 1xnb with bootstrap replicas of the quantile.

References

Cuevas A, Febrero M, Fraiman R. 2006. On the use of bootstrap for estimating functions with functional data. Computational Statistics and Data Analysis 51: 1063-1074.

Febrero-Bande, M., Galeano, P., and Gonzalez-Manteiga, W. (2008). Outlier detection in functional data by depth measures with application to identify abnormal NOx levels. Environmetrics 19, 4, 331–345.

Febrero-Bande, M., Galeano, P. and Gonzalez-Manteiga, W. (2007). A functional analysis of NOx levels: location and scale estimation and outlier detection. Computational Statistics 22, 3, 411-427.

Febrero-Bande, M., Oviedo de la Fuente, M. (2012). Statistical Computing in Functional Data Analysis: The R Package fda.usc. Journal of Statistical Software, 51(4), 1-28. https://www.jstatsoft.org/v51/i04/

See also

Author

Manuel Febrero-Bande, Manuel Oviedo de la Fuente manuel.oviedo@udc.es

Examples

if (FALSE) { # \dontrun{
data(aemet)
nb=20 # Time consuming
out.trim<-outliers.depth.trim(aemet$temp,dfunc=depth.FM,nb=nb)
plot(aemet$temp,col=1,lty=1)
lines(aemet$temp[out.trim[[1]]],col=2)
} # }