PCvM statistic for the Functional Linear Model with scalar response

Projected Cramer-von Mises statistic (PCvM) for the Functional Linear Model with scalar response (FLM): \(Y=\big<X,\beta\big>+\varepsilon\).

Adot(X, inpr)

PCvM.statistic(X, residuals, p, Adot.vec)

Arguments

X

Functional covariate for the FLM. The object must be either in the class fdata or in the class fd. It is used to compute the matrix of inner products.

inpr

Matrix of inner products of X. Computed if not given.

residuals

Residuals of the estimated FLM.

p

Number of elements of the functional basis where the functional covariate is represented.

Adot.vec

Output from the Adot function (see Details). Computed if not given.

Value

For PCvM.statistic, the value of the statistic. For Adot, a suitable output to be used in the argument Adot.vec.

Details

In order to optimize the computation of the statistic, the critical parts of these two functions are coded in FORTRAN. The hardest part corresponds to the function Adot, which involves the computation of a symmetric matrix of dimension \(n\times n\) where each entry is a sum of \(n\) elements. As this matrix is symmetric, the order of the method can be reduced from \(O(n^3)\) to \(O\big(\frac{n^3-n^2}{2}\big)\). The memory requirement can also be reduced to \(O\big(\frac{n^2-n+2}{2}\big)\). The value of Adot is a vector of length \(\frac{n^2-n+2}{2}\) where the first element is the common diagonal element and the rest are the lower triangle entries of the matrix, sorted by rows (see Examples).

Note

No NA's are allowed in the functional covariate.

References

Escanciano, J. C. (2006). A consistent diagnostic test for regression models using projections. Econometric Theory, 22, 1030-1051. doi:10.1017/S0266466606060506

Garcia-Portugues, E., Gonzalez-Manteiga, W. and Febrero-Bande, M. (2014). A goodness--of--fit test for the functional linear model with scalar response. Journal of Computational and Graphical Statistics, 23(3), 761-778. doi:10.1080/10618600.2013.812519

See also

Author

Eduardo Garcia-Portugues. Please, report bugs and suggestions to eduardo.garcia.portugues@uc3m.es

Examples

# Functional process
X=rproc2fdata(n=10,t=seq(0,1,l=101))
# Adot
Adot.vec=Adot(X)

# Obtain the entire matrix Adot
Ad=diag(rep(Adot.vec[1],dim(X$data)[1]))
Ad[upper.tri(Ad,diag=FALSE)]=Adot.vec[-1]
Ad=t(Ad)
Ad=Ad+t(Ad)-diag(diag(Ad))
Ad
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 34.55752 22.51097 22.68963 22.86566 23.36413 22.86389 22.23561 24.19536
#>  [2,] 22.51097 34.55752 22.46356 23.36041 22.86665 22.84161 23.50343 23.66540
#>  [3,] 22.68963 22.46356 34.55752 22.15425 23.43818 23.67042 23.37163 22.05998
#>  [4,] 22.86566 23.36041 22.15425 34.55752 22.48090 22.94446 23.53226 24.07350
#>  [5,] 23.36413 22.86665 23.43818 22.48090 34.55752 23.72822 23.70668 22.84141
#>  [6,] 22.86389 22.84161 23.67042 22.94446 23.72822 34.55752 23.39016 22.86442
#>  [7,] 22.23561 23.50343 23.37163 23.53226 23.70668 23.39016 34.55752 22.87583
#>  [8,] 24.19536 23.66540 22.05998 24.07350 22.84141 22.86442 22.87583 34.55752
#>  [9,] 22.45255 21.84645 22.39000 23.38538 23.04749 23.57639 22.20286 23.96308
#> [10,] 23.31920 23.03904 21.98842 23.26944 23.51096 23.25699 22.99696 22.71862
#>           [,9]    [,10]
#>  [1,] 22.45255 23.31920
#>  [2,] 21.84645 23.03904
#>  [3,] 22.39000 21.98842
#>  [4,] 23.38538 23.26944
#>  [5,] 23.04749 23.51096
#>  [6,] 23.57639 23.25699
#>  [7,] 22.20286 22.99696
#>  [8,] 23.96308 22.71862
#>  [9,] 34.55752 23.20319
#> [10,] 23.20319 34.55752
# Statistic
PCvM.statistic(X,residuals=rnorm(10),p=5)
#> [1] 3.305956