Kernel Classifier from Functional Data

Fits Nonparametric Supervised Classification for Functional Data.

Usage

classif.np(
  group,
  fdataobj,
  h = NULL,
  Ker = AKer.norm,
  metric,
  weights = "equal",
  type.S = S.NW,
  par.S = list(),
  ...
)

classif.knn(
  group,
  fdataobj,
  knn = NULL,
  metric,
  weights = "equal",
  par.S = list(),
  ...
)

classif.kernel(
  group,
  fdataobj,
  h = NULL,
  Ker = AKer.norm,
  metric,
  weights = "equal",
  par.S = list(),
  ...
)

Arguments

group

Factor of length n

fdataobj

fdata class object.

h

Vector of smoothing parameter or bandwidth.

Ker

Type of kernel used.

metric

Metric function, by default metric.lp.

weights

weights.

type.S

Type of smothing matrix S. By default S is calculated by Nadaraya-Watson kernel estimator (S.NW).

par.S

List of parameters for type.S: w, the weights.

...

Arguments to be passed for metric.lp o other metric function and Kernel function.

knn

Vector of number of nearest neighbors considered.

Value

fdataobj: fdata class object.

group: Factor of length n.

group.est: Estimated vector groups.

prob.group: Matrix of predicted class probabilities. For each functional point shows the probability of each possible group membership.

max.prob: Highest probability of correct classification.

h.opt: Optimal smoothing parameter or bandwidht estimated.

D: Matrix of distances of the optimal quantile distance hh.opt.

prob.classification: Probability of correct classification by group.

misclassification: Vector of probability of misclassification by number of neighbors knn.

h: Vector of smoothing parameter or bandwidht.

C: A call of function classif.kernel.

Details

Make the group classification of a training dataset using kernel or KNN estimation: Kernel.

Different types of metric funtions can be used.

Note

If fdataobj is a data.frame the function considers the case of multivariate covariates.
metric.dist function is used to compute the distances between the rows of a data matrix (as dist function.

References

Ferraty, F. and Vieu, P. (2006). Nonparametric functional data analysis. Springer Series in Statistics, New York.

Ferraty, F. and Vieu, P. (2006). NPFDA in practice. Free access on line at https://www.math.univ-toulouse.fr/~ferraty/SOFTWARES/NPFDA/

See also

Author

Manuel Febrero-Bande, Manuel Oviedo de la Fuente manuel.oviedo@udc.es

Examples

if (FALSE) { # \dontrun{
data(phoneme)
mlearn <- phoneme[["learn"]]
glearn <- phoneme[["classlearn"]]
h <- 9:19
out <- classif.np(glearn,mlearn,h=h)
summary(out)
head(round(out$prob.group,4))
} # }