Proximities between functional data

Computes the cosine correlation distance between two functional dataset of class fdata.

Usage

dis.cos.cor(fdata1, fdata2 = NULL, as.dis = FALSE)

Arguments

fdata1

Functional data 1 or curve 1.

fdata2

Functional data 2 or curve 2.

as.dis

Returns the distance matrix como 1-cor(fdata1,fdata2).

Value

Returns a proximity/distance matrix (depending on as.dis) between functional data.

References

Kemmeren P, van Berkum NL, Vilo J, et al. (2002). Protein Interaction Verification and Functional Annotation by Integrated Analysis of Genome-Scale Data . Mol Cell. 2002 9(5):1133-43.

See also

Examples

if (FALSE) { # \dontrun{
 r1<-rnorm(1001,sd=.01)
 r2<-rnorm(1001,sd=.01)
 x<-seq(0,2*pi,length=1001)
 fx<-fdata(sin(x)/sqrt(pi)+r1,x)
 dis.cos.cor(fx,fx)
 dis.cos.cor(c(fx,fx),as.dis=TRUE)
 fx0<-fdata(rep(0,length(x))+r2,x)
 plot(c(fx,fx0))
 dis.cos.cor(c(fx,fx0),as.dis=TRUE)
 } # }