Calculation of the smoothing parameter (h) for a functional data

Calculation of the smoothing parameter (h) for a functional data using nonparametric kernel estimation.

Usage

h.default(
  fdataobj,
  prob = c(0.025, 0.25),
  len = 51,
  metric = metric.lp,
  type.S = "S.NW",
  Ker = Ker.norm,
  ...
)

Arguments

fdataobj

fdata class object.

prob

Vector of probabilities for extracting the quantiles of the distance matrix. If length(prob)=2 a sequence between prob[1] and prob[2] of length len.

len

Vector length of smoothing parameter h to return only used
when length(prob)=2.

metric

If is a function: name of the function to calculate the distance matrix between the curves, by default metric.lp. If is a matrix: distance matrix between the curves. kernel.

type.S

Type of smothing matrix S. Possible values are: Nadaraya-Watson estimator "S.NW" and K nearest neighbors estimator "S.KNN"

Ker

Kernel function. By default, Ker.norm. Useful for scaling the bandwidth values according to Kernel

...

Arguments to be passed for metric argument.

Value

Returns the vector of smoothing parameter or bandwidth h.

See also

Author

Manuel Febrero-Bande, Manuel Oviedo de la Fuente manuel.oviedo@udc.es

Examples

if (FALSE) { # \dontrun{
data(aemet)
h1<-h.default(aemet$temp,prob=c(0.025, 0.25),len=2)
mdist<-metric.lp(aemet$temp)
h2<-h.default(aemet$temp,len=2,metric=mdist)
h3<-h.default(aemet$temp,len=2,metric=semimetric.pca,q=2)
h4<-h.default(aemet$temp,len=2,metric=semimetric.pca,q=4)
h5<-h.default(aemet$temp,prob=c(.2),type.S="S.KNN")
h1;h2;h3;h4;h5
} # }