Summarizes information from kernel classification methods.

Summary function for classif.knn or classif.kernel.

Usage

# S3 method for class 'classif'
summary(object, ...)

# S3 method for class 'classif'
print(x, digits = max(3, getOption("digits") - 3), ...)

Arguments

object

Estimated by kernel classification.

...

Further arguments passed to or from other methods.

x

Estimated by kernel classification.

digits

how many significant digits are to be used for numeric and complex x.

Value

Shows:

-Probability of correct classification by group prob.classification.

-Confusion matrix between the theoretical groups and estimated groups.

-Highest probability of correct classification max.prob.

If the object is returned from the function classif.knn

-Vector of probability of correct classification by number of neighbors knn.

-Optimal number of neighbors: knn.opt.

If the object is returned from the function: classif.kernel

-Vector of probability of correct classification by banwidth h.

-Functional measure of closeness (optimal distance, h.opt).

Details

object from classif.knn or classif.kernel

See also

Author

Manuel Febrero-Bande, Manuel Oviedo de la Fuente manuel.oviedo@udc.es

Examples

if (FALSE) { # \dontrun{ 
data(phoneme)
mlearn<-phoneme[["learn"]]
glearn<-phoneme[["classlearn"]]
out=classif.knn(glearn,mlearn,knn=c(3,5,7))
summary(out)
out2=classif.kernel(glearn,mlearn,h=2^(0:5))
summary(out2)
} # }